Eliminated overhead of static transparency animations
All or nearly all M2s ship with transparency animation blocks, even
if there is no transparency animation used in the M2. In cases where
there is no animation, the block contains a single track with a single
value: 1.0 (in fixed precision).

To reduce the overhead of spawning animation tracks in code that relies
on Blizzardry, some additional helpers are now utilized to prevent
flagging the M2 as animated in this particular case.